refactor: decompose - create textArea component
